fix discount validation to check max item discount first
This commit fixes the issue where the condition to check for the max discount percentage from the pos profile settings would contradict with item max discount percentage condition.

now, it checks if an item has max discount set and validates against that, if the conditions are not met and the item does not have a max discount value set then it validates against the pos profile max discount field if set.