Feat: Round to nearest multiple of five
Round to nearest multiple of five:

New Option in POSAwesome Settings:  Round Discounts to nearest five

if this new option is enabled it will:

round down any item.rate to nearest multiple of 5.
round total and subtotal of an invoice to nearest multiple of 5.
it will apply the appropriate values to: (additional_discount_percent, discount_amount, item.discount_amount, and item.discount_percentage)

if you have item groups  that are not priced according to multiple of 5 for example an item with a price_list as 3.5$, you can add them to (Don't round for item groups.) multi Select field and it will skip rounding for items related to these item groups. however it will still round the invoice total and subtotal.